chore: restructure monorepo
