+++ packages/core/package.json
@@ -1,0 +1,50 @@
+{
+	"name": "@sheepdog/core",
+	"version": "0.0.1",
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/mainmatter/sheepdog.git"
+	},
+	"main": "./dist/index.js",
+	"type": "module",
+	"scripts": {
+		"build": "tsc --outDir dist --declaration --sourceMap false && publint",
+		"prepack": "pnpm build",
+		"prepare": "pnpm build"
+	},
+	"files": [
+		"dist"
+	],
+	"exports": {
+		".": {
+			"types": "./dist/index.d.ts",
+			"svelte": "./dist/index.js",
+			"import": "./dist/index.js"
+		},
+		"./utils": {
+			"types": "./dist/utils.d.ts",
+			"svelte": "./dist/utils.js",
+			"import": "./dist/utils.js"
+		},
+		"./vite": {
+			"types": "./dist/vite.d.ts",
+			"svelte": "./dist/vite.js",
+			"import": "./dist/vite.js"
+		}
+	},
+	"types": "./dist/index.d.ts",
+	"svelte": "./dist/index.js",
+	"dependencies": {
+		"acorn": "^8.11.3",
+		"esm-env": "^1.1.4",
+		"esrap": "^1.2.2",
+		"zimmerframe": "^1.1.2"
+	},
+	"volta": {
+		"extends": "../../package.json"
+	},
+	"devDependencies": {
+		"publint": "^0.2.7",
+		"vite": "^5.2.11"
+	}
+}
