+++ packages/core/tsconfig.json
@@ -1,0 +1,6 @@
+{
+	"extends": "../../tsconfig.base.json",
+	"compilerOptions": {
+		"composite": true
+	}
+}
