+++ tsconfig.json
@@ -1,3 +1,3 @@
 {
-	"extends": "./tsconfig.base.json"
+	"references": [{ "path": "./packages/core" }]
 }
